refactor: relocated code
