+++ .github/workflows/ci.yml
@@ -34,6 +34,7 @@ jobs:
         python${{ env.PYTHON_VERSION }} -m venv .venv
         source .venv/bin/activate
         make install-ci
+        
   lint:
     runs-on: ubuntu-latest
     needs: setup
